Remove linear factor from Engine's consume (#494)
20x perf speedup in the "BasicBacktrack" benchmarks.
